web/src/lib/auth/csrf.service.ts
import { CsrfToken, CsrfConfig, CsrfService } from './csrf.types';
import crypto from 'crypto';

export class CsrfServiceImpl implements CsrfService {
  private config: CsrfConfig = {
    headerName: 'X-CSRF-Token',
    cookieName: 'csrf_token',
    ignoredMethods: ['GET', 'HEAD', 'OPTIONS'],
    tokenLength: 32,
    expiryTime: 3600
  };

  private generateSecureToken(): string {
    return crypto.randomBytes(this.config.tokenLength).toString('hex');
  }

  async generateToken(sessionId: string): Promise<CsrfToken> {
    if (!sessionId) {
      throw new Error('Session ID is required');
    }

    const now = Math.floor(Date.now() / 1000);
    return {
      token: this.generateSecureToken(),
      sessionId,
      issuedAt: now,
      expiresAt: now + this.config.expiryTime
    };
  }

  async validateToken(sessionId: string, token: string): Promise<boolean> {
    if (!sessionId || !token) {
      return false;
    }

    try {
      // In a real implementation, you would validate against stored tokens
      // For now, we'll just check if the token exists and isn't expired
      const now = Math.floor(Date.now() / 1000);
      const storedToken = await this.getStoredToken(sessionId);
      
      if (!storedToken) {
        return false;
      }

      return (
        storedToken.token === token &&
        storedToken.sessionId === sessionId &&
        storedToken.expiresAt > now
      );
    } catch (error) {
      console.error('Error validating CSRF token:', error);
      return false;
    }
  }

  async refreshToken(oldToken: CsrfToken): Promise<CsrfToken> {
    if (!oldToken || !oldToken.sessionId) {
      throw new Error('Invalid token provided');
    }

    const now = Math.floor(Date.now() / 1000);
    if (now > oldToken.expiresAt) {
      throw new Error('Token has expired');
    }

    // Generate a new token with the same session ID
    return this.generateToken(oldToken.sessionId);
  }

  // This is a placeholder method - in a real implementation,
  // you would interact with your session storage
  private async getStoredToken(sessionId: string): Promise<CsrfToken | null> {
    // TODO: Implement actual token storage and retrieval
    // This would typically involve your session storage mechanism
    return null;
  }
}
